pyproject.toml
[project]
name = "finback"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django>=5.1.4",
    "pillow>=11.1.0",
    "pyecharts>=2.0.7",
]
